requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
ipykernel==6.29.5
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyzmq==27.0.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
traitlets==5.14.3
typing-extensions==4.14.0
wcwidth==0.2.13
gradio
